+++ resources/providers/config.rb
@@ -6,21 +6,6 @@ action :add do
     sensor_id = new_resource.sensor_id
     groups = new_resource.groups
 
-    # temporary until we configure selinux
-    ruby_block 'set_enforce_0' do
-      block do
-        selinux_status = `getenforce`
-
-        if selinux_status.strip != 'Permissive'
-          Chef::Log.warn('Setting SELinux to permissive mode...')
-          system('setenforce 0')
-        else
-          Chef::Log.info('SELinux is already in permissive mode.')
-        end
-      end
-      action :run
-    end
-
     ruby_block 'check_bpctl_mod' do
       block do
         module_loaded = `lsmod | grep bpctl_mod`
@@ -37,29 +22,50 @@ action :add do
 
     dnf_package 'snort3' do
       action :upgrade
-      flush_cache [:before]
     end
 
     valid_instance_names = []
 
     groups.each do |group|
-      name = group['name']
+      group_name = group['name']
+      default_added = false
+      group['bindings']
+        .keys
+        .map(&:to_i)
+        .sort
+        .map(&:to_s)
+        .each do |id_str|
+        vgroup         = group['bindings'][id_str].to_hash.clone
+        vgroup_name    = vgroup['name'].nil? ? 'default' : vgroup['name'].to_s
+        vgroup['name'] = vgroup_name
+        binding_id     = id_str.to_i
+        vgroup['id']   = binding_id
 
-      bindings = group['bindings'].keys.map(&:to_i).sort
+        has_vlans   = vgroup['vlan_objects'] && !vgroup['vlan_objects'].empty?
+        has_network = vgroup['network_objects'] && !vgroup['network_objects'].empty?
 
-      bindings.each do |binding_id|
-        vgroup = {}
-        vgroup['ipvars']        = node['redborder']['snort']['default']['ipvars']
-        vgroup['portvars']      = node['redborder']['snort']['default']['portvars']
+        if !has_vlans && !has_network
+          if default_added
+            next
+          else
+            default_added = true
+          end
+        end
+
+        if vgroup['ipvars'].nil? || vgroup['ipvars'].empty?
+          vgroup['ipvars'] = node['redborder']['snort']['default']['ipvars']
+        end
+
+        if vgroup['portvars'].nil? || vgroup['portvars'].empty?
+          vgroup['portvars'] = node['redborder']['snort']['default']['portvars']
+        end
 
         instance_name = "#{group['instances_group']}_#{group['name']}_#{binding_id}"
-
         valid_instance_names << "snort3@#{instance_name}.service"
 
         service "snort3@#{instance_name}.service" do
           action [:enable]
         end
-
         %w(reload restart stop start).each do |s_action|
           execute "#{s_action}_snort3@#{instance_name}" do
             command "/bin/env WAIT=1 /bin/systemctl #{s_action} snort3@#{instance_name}.service"
@@ -84,6 +90,86 @@ action :add do
           action :create
         end
 
+        ruby_block 'copy_alerts_with_full_date_hour_shell' do
+          block do
+            timestamp = Time.now.strftime('%Y-%m-%d_%H')
+            src_dir = '/etc/snort/0_default_0'
+            raw_dir = "#{src_dir}/raw"
+            dest_dir = "#{raw_dir}/#{timestamp}"
+
+            system(<<-EOS
+              mkdir -p "#{dest_dir}"
+              cd "#{src_dir}" || exit 1
+
+              for file in *_alert_full.txt; do
+                [ -e "$file" ] || continue
+
+                base_name="$(basename "$file")"
+                name="${base_name%.*}"
+                ext="${base_name##*.}"
+
+                dest_file="#{dest_dir}/$base_name"
+
+                if [ -e "$dest_file" ]; then
+                  i=1
+                  while [ -e "#{dest_dir}/$name_$i.$ext" ]; do
+                    i=$((i + 1))
+                  done
+                  dest_file="#{dest_dir}/$name_$i.$ext"
+                fi
+
+                cp -f "$file" "$dest_file"
+                : > "$file"
+                echo "Copied $file to $dest_file and truncated original"
+              done
+
+              find "#{dest_dir}" -type f -size 0 -name '*.txt' -delete
+            EOS
+                  )
+          end
+          action :run
+        end
+
+        ruby_block 'copy_alerts_with_full_date_hour_shell' do
+          block do
+            timestamp = Time.now.strftime('%Y-%m-%d_%H')
+            src_dir = '/etc/snort/0_default_0'
+            raw_dir = "#{src_dir}/raw"
+            dest_dir = "#{raw_dir}/#{timestamp}"
+
+            system(<<-EOS
+              mkdir -p "#{dest_dir}"
+              cd "#{src_dir}" || exit 1
+
+              for file in *_alert_full.txt; do
+                [ -e "$file" ] || continue
+
+                base_name="$(basename "$file")"
+                name="${base_name%.*}"
+                ext="${base_name##*.}"
+
+                dest_file="#{dest_dir}/$base_name"
+
+                if [ -e "$dest_file" ]; then
+                  i=1
+                  while [ -e "#{dest_dir}/$name_$i.$ext" ]; do
+                    i=$((i + 1))
+                  done
+                  dest_file="#{dest_dir}/$name_$i.$ext"
+                fi
+
+                cp -f "$file" "$dest_file"
+                : > "$file"
+                echo "Copied $file to $dest_file and truncated original"
+              done
+
+              find "#{dest_dir}" -type f -size 0 -name '*.txt' -delete
+            EOS
+                  )
+          end
+          action :run
+        end
+
         begin
           sensor_id = node['redborder']['sensor_id'].to_i
         rescue
@@ -97,14 +183,14 @@ action :add do
           group 'root'
           mode '0644'
           retries 2
-          variables(instance_name: instance_name, group: group, sensor_id: sensor_id, name: name)
+          variables(instance_name: instance_name, group: group, sensor_id: sensor_id, group_name: group_name)
           notifies :stop, "service[snort3@#{instance_name}.service]", :delayed
           notifies :start, "service[snort3@#{instance_name}.service]", :delayed
         end
 
         segment = group['segments'].join(' ')
         iface = `ip link show master #{group['segments'].join(' ')} | grep '^[0-9]' | awk '{print $2}' | cut -d':' -f1 | paste -sd ":"`.chomp!
-        threads = group['cpu_list'].size * 2
+        threads = group['cpu_list'].size
         cpu_cores = group['cpu_list'].join(' ')
         mode = group['mode']
         inline = (mode != 'IDS' && mode != 'IDS_SPAN') && (mode == 'IPS' || mode == 'IDS_FWD' || mode == 'IPS_TEST')
@@ -181,6 +267,53 @@ action :add do
             end
           end
           only_if { ::File.exist?("/etc/snort/#{instance_name}/env") }
+        end
+  
+        template "/etc/snort/#{instance_name}/snort.rules" do
+          source 'empty.erb'
+          cookbook 'snort3'
+          owner 'root'
+          group 'root'
+          mode '0644'
+          retries 2
+          not_if { ::File.exist?("/etc/snort/#{instance_name}/snort.rules") && !::File.zero?("/etc/snort/#{instance_name}/snort.rules") }
+        end
+
+        template "/etc/snort/#{instance_name}/events.lua" do
+          source 'empty.erb'
+          cookbook 'snort3'
+          owner 'root'
+          group 'root'
+          mode '0644'
+          retries 2
+          not_if { ::File.exist?("/etc/snort/#{instance_name}/events.lua") && !::File.zero?("/etc/snort/#{instance_name}/events.lua") }
+        end
+
+        template_paths = {
+          'iplists/allowlist'    => "/etc/snort/#{instance_name}/iplists",
+          'iplists/blacklist'    => "/etc/snort/#{instance_name}/iplists",
+          'iplists/monitorlist'  => "/etc/snort/#{instance_name}/iplists",
+          'geoips/rbgeoip'       => "/etc/snort/#{instance_name}/geoips",
+        }
+
+        template_paths.each do |relative_path, dir_path|
+          directory dir_path do
+            recursive true
+            owner 'root'
+            group 'root'
+            mode '0755'
+            action :create
+          end
+
+          template "/etc/snort/#{instance_name}/#{relative_path}" do
+            source 'empty.erb'
+            cookbook 'snort3'
+            owner 'root'
+            group 'root'
+            mode '0644'
+            retries 2
+            action :create_if_missing
+          end
         end
 
         template "/etc/snort/#{instance_name}/snort-variables.conf" do