+++ CHANGELOG.md
@@ -1,6 +1,44 @@
 cookbook-snort3 CHANGELOG
 ===============
 
+## 0.1.7
+
+  - Miguel Alvarez
+    - [6f279f7] Remove force-setenforce 0 (#12)
+    - [280c06a] Add intrusion thresholds (#11)
+
+## 0.1.6
+
+  - Miguel Alvarez
+    - [280c06a] Add intrusion thresholds (#11)
+
+## 0.1.5
+
+  - nilsver
+    - [d3c7d69] remove flush cache
+
+## 0.1.4
+
+  - Miguel Alvarez
+    - [a434c47] Intrusion snort3 reputation (#6)
+    - [0b17e04] Create snort.rules file for intrusion (#3)
+
+## 0.1.3
+
+  - Miguel Alvarez
+    - [5c0c817] Add raw view for intrusion sensor (#5)
+    - [0b17e04] Create snort.rules file for intrusion (#3)
+
+## 0.1.2
+
+  - Miguel Alvarez
+    - [0b17e04] Create snort.rules file for intrusion (#3)
+
+## 0.1.1
+
+  - Miguel Alvarez
+    - [262b1b8] IPvars per group and not global in intrusion
+
 ## 0.1.0
 
   - Miguel Alvarez